+++ demo/storage/index.ts
@@ -9,46 +9,48 @@ interface User extends DocumentObject {
   token?: string;
 }
 
-const db = new AlwatrStorage<User>({name: 'user-list', path: 'db'});
+const db = new AlwatrStorage<User>({
+  name: 'user-list',
+  path: 'db',
+  saveBeautiful: true,
+  debug: true,
+});
 
-// await db.readyPromise
-// or
-db.readyPromise.then(() => {
-  console.log('db loaded and ready to access.');
+console.log('db loaded and ready to access.');
 
-  let ali = db.get('alimd');
+let ali = db.get('alimd');
 
-  if (ali == null) {
-    console.log('ali not found');
-    ali = {
-      _id: 'alimd',
-      _updatedBy: 'demo',
-      fname: 'Ali',
-      lname: 'Mihandoost',
-      email: 'ali@mihandoost.com',
-    };
-  } else {
-    console.log('ali found: %o', ali);
-    /**
-     * {
-     *   _id: 'alimd',
-     *   fname: 'Ali',
-     *   lname: 'MM',
-     *   email: 'i@ali.md',
-     * }
-     */
-
-    ali.token = Math.random().toString(36).substring(2, 15);
-  }
-
-  db.set(ali);
-
-  db.set({
-    _id: 'fmd',
+if (ali == null) {
+  console.log('ali not found');
+  ali = {
+    _id: 'alimd',
     _updatedBy: 'demo',
-    fname: 'Fatemeh',
+    fname: 'Ali',
     lname: 'Mihandoost',
-    email: 'Fatemeh@mihandoost.com',
-    token: Math.random().toString(36).substring(2, 15),
-  });
+    email: 'ali@mihandoost.com',
+  };
+}
+else {
+  console.log('ali found: %o', ali);
+  /**
+   * {
+   *   _id: 'alimd',
+   *   fname: 'Ali',
+   *   lname: 'MM',
+   *   email: 'i@ali.md',
+   * }
+   */
+
+  ali.token = Math.random().toString(36).substring(2, 15);
+}
+
+db.set(ali);
+
+db.set({
+  _id: 'fmd',
+  _updatedBy: 'demo',
+  fname: 'Fatemeh',
+  lname: 'Mihandoost',
+  email: 'Fatemeh@mihandoost.com',
+  token: Math.random().toString(36).substring(2, 15),
 });