+++ packages/core/nano-server/src/nano-server.ts
@@ -183,10 +183,12 @@ export class AlwatrNanoServer {
     try {
       if (typeof middleware === 'function') {
         await middleware(connection);
-      } else {
+      }
+      else {
         this._notFoundListener(connection);
       }
-    } catch (err) {
+    }
+    catch (err) {
       this._logger.error('handleRequest', 'http_server_middleware_error', err, {
         method: connection.method,
         route,
@@ -276,16 +278,17 @@ export class AlwatrConnection {
     let body = '';
     try {
       body = JSON.stringify(content);
-    } catch {
+    }
+    catch {
       this._logger.accident('responseData', 'data_stringify_failed', 'JSON.stringify(data) failed!');
       return this.reply(
-        content.ok === false ?
-          {
+        content.ok === false
+          ? {
             ok: false,
             statusCode: content.statusCode,
             errorCode: content.errorCode,
-          } :
-          {
+          }
+          : {
             ok: false,
             statusCode: 500,
             errorCode: 'data_stringify_failed',
@@ -362,7 +365,8 @@ export class AlwatrConnection {
 
     try {
       return JSON.parse(body) as Type;
-    } catch (err) {
+    }
+    catch (err) {
       this.reply({
         ok: false,
         statusCode: 400,