+++ packages/core/signal/src/core.ts
@@ -73,7 +73,8 @@ function __callListeners<SignalName extends keyof AlwatrSignals>(signal: SignalO
           }),
         );
       }
-    } catch (err) {
+    }
+    catch (err) {
       logger.error('_callListeners', 'call_listener_failed', (err as Error).stack || err, {
         signalName: signal.name,
       });
@@ -98,14 +99,19 @@ function __callListeners<SignalName extends keyof AlwatrSignals>(signal: SignalO
 export function _addSignalListener<SignalName extends keyof AlwatrSignals>(
     signal: SignalObject<SignalName>,
     listenerCallback: ListenerCallback<SignalName>,
-    options?: ListenerOptions,
+    options: ListenerOptions = {},
 ): ListenerObject<SignalName> {
+  options.once ??= false;
+  options.disabled ??= false;
+  options.receivePrevious ??= true;
+  options.priority ??= false;
+
   logger.logMethodArgs('_addSignalListener', {signalName: signal.name, options});
 
   const listener: ListenerObject<SignalName> = {
     id: ++_lastListenerId,
-    once: options?.once ?? false,
-    disabled: options?.disabled ?? false,
+    once: options.once,
+    disabled: options.disabled,
     callback: listenerCallback,
   };
 
@@ -114,24 +120,31 @@ export function _addSignalListener<SignalName extends keyof AlwatrSignals>(
   // Run callback for old dispatch signal
   if (signal.value !== undefined) {
     // null is a valid value for signal.
-    if (options?.receivePrevious === 'Immediate') {
-      logger.incident('_addSignalListener', 'call_signal_callback', 'run callback with previous signal value!', {
-        signalName: signal.name,
-        mode: 'Immediate',
-      });
+    if (options.receivePrevious === 'Immediate') {
+      logger.incident(
+          '_addSignalListener',
+          'call_signal_callback',
+          'run callback with previous signal value',
+          {
+            signalName: signal.name,
+            mode: 'Immediate',
+          },
+      );
       try {
         listenerCallback(signal.value);
-      } catch (err) {
+      }
+      catch (err) {
         logger.error('_addSignalListener', 'call_signal_callback_failed', (err as Error).stack || err, {
           signalName: signal.name,
         });
       }
       callbackCalled = true;
-    } else if (options?.receivePrevious === true) {
+    }
+    else if (options.receivePrevious === true) {
       requestAnimationFrame(() => {
         if (signal.value !== undefined) {
           // null is a valid value for signal.
-          logger.incident('_addSignalListener', 'call_signal_callback', 'run callback with previous signal value!', {
+          logger.incident('_addSignalListener', 'call_signal_callback', 'run callback with previous signal value', {
             signalName: signal.name,
             mode: 'Delay',
           });
@@ -143,10 +156,11 @@ export function _addSignalListener<SignalName extends keyof AlwatrSignals>(
   }
 
   // if once then must remove listener after fist callback called! then why push it to listenerList?!
-  if (!(options?.once && callbackCalled)) {
-    if (options?.priority) {
+  if (!(options.once === true && callbackCalled === true)) {
+    if (options.priority === true) {
       signal.listenerList.unshift(listener);
-    } else {
+    }
+    else {
       signal.listenerList.push(listener);
     }
   }
@@ -186,17 +200,19 @@ export function _removeSignalListener<SignalName extends keyof AlwatrSignals>(
 export function _dispatchSignal<SignalName extends keyof AlwatrSignals>(
     signal: SignalObject<SignalName>,
     value: AlwatrSignals[SignalName],
-    options?: DispatchOptions,
+    options: DispatchOptions = {},
 ): void {
+  options.debounce ??= true;
+
   logger.logMethodArgs('dispatchSignal', {signalName: signal.name, value, options});
 
   // set value before check signal.debounced for act like throttle (call listeners with last dispatch value).
   signal.value = value;
 
   if (signal.disabled) return; // signal is disabled.
-  if (options?.debounce && signal.debounced) return; // last dispatch in progress.
+  if (options.debounce === true && signal.debounced === true) return; // last dispatch in progress.
 
-  if (!options?.debounce) {
+  if (options.debounce !== true) {
     // call listeners immediately.
     __callListeners(signal);
     return;
@@ -232,8 +248,11 @@ export function _setSignalProvider<SignalName extends keyof AlwatrSignals>(
     signal: SignalObject<SignalName>,
     requestSignal: SignalObject<SignalName>,
     signalProvider: SignalProvider<SignalName>,
-    options?: SignalProviderOptions,
+    options: SignalProviderOptions = {},
 ): ListenerObject<SignalName> {
+  options.debounce ??= true;
+  options.receivePrevious ??= true;
+
   logger.logMethodArgs('_setSignalProvider', {signal: signal.name, requestSignal: requestSignal.name, options});
 
   if (requestSignal.listenerList.length > 0) {
@@ -252,11 +271,11 @@ export function _setSignalProvider<SignalName extends keyof AlwatrSignals>(
     const signalValue = await signalProvider(requestParam);
     if (signalValue !== undefined) {
       // null is a valid value for signal.
-      _dispatchSignal(signal, signalValue, {debounce: options?.debounce ?? true});
+      _dispatchSignal(signal, signalValue, {debounce: options.debounce});
     }
   };
 
   return _addSignalListener(requestSignal, _callback as unknown as ListenerCallback<SignalName>, {
-    receivePrevious: options?.receivePrevious ?? true,
+    receivePrevious: options.receivePrevious,
   });
 }